+++ src/tableaus/tableaus_iprk.jl
@@ -1,0 +1,6 @@
+
+"Gauss-Legendre Runge-Kutta"
+function getTableauIPGLRK(s::Int)
+    TableauIPRK(Symbol("IPGLRK", s), 2^s, TableauFIRK(getCoefficientsGLRK(s)))
+    
+end