+++ docs/src/releasenotes.md
@@ -7,6 +7,7 @@
 
 * Atomic solutions can now store a NamedTuple of internal variables of the integrator, including nonlinear solver output
 * Output of internal variables has been added to VPRK integrators
+* Add Gauss-Legendre tableaus for implicit partitioned Runge-Kutta methods
 
 ### Fixes